refactor: clean up main server logic
